# jws: improve performance and allocations for ParseSignedCompact