--- a/jwe_test.go
+++ b/jwe_test.go
@@ -717,3 +717,12 @@ func TestJWEWithNullAlg(t *testing.T) {
 		t.Error(err)
 	}
 }
+
+func BenchmarkParseEncryptedCompat(b *testing.B) {
+	msg := "eyJhbGciOiJSU0EtT0FFUCIsImVuYyI6IkExMjhHQ00ifQ.dGVzdA.dGVzdA.dGVzdA.dGVzdA"
+	for b.Loop() {
+		if _, err := ParseEncryptedCompact(msg, []KeyAlgorithm{RSA_OAEP}, []ContentEncryption{A128GCM}); err != nil {
+			panic(err)
+		}
+	}
+}