See also https://go-review.googlesource.com/c/oauth2/+/655455

name                  old time/op    new time/op    delta
ParseSignedCompat-16    1.95µs ± 1%    1.91µs ± 1%  -1.83%  (p=0.000 n=9+10)

name                  old alloc/op   new alloc/op   delta
ParseSignedCompat-16    3.26kB ± 0%    3.21kB ± 0%  -1.47%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
ParseSignedCompat-16      45.0 ± 0%      44.0 ± 0%  -2.22%  (p=0.000 n=10+10)